Move UI code to common directory.
